requirements.txt
amqp==5.2.0
antiorm==1.2.1
awscrt==0.21.1
AWSIoTPythonSDK==1.5.4
billiard==4.2.1
boto3==1.35.33
botocore==1.35.33
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorzero==2.0
db==0.1.1
db-sqlite3==0.0.1
gpiozero==2.0.1
idna==3.10
jmespath==1.0.1
keyboard==0.13.5
kombu==5.3.4
numpy==2.1.1
paho-mqtt==2.1.0
pillow==10.4.0
prompt_toolkit==3.0.48
python-dateutil
python-dotenv
redis==3.5.3
requests==2.32.3
RPi.GPIO==0.7.1
s3transfer==0.10.2
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
